Move WebGL layer into its own file
