+++ src/crafty.js
@@ -33,6 +33,7 @@ require('./graphics/sprite');
 require('./graphics/text');
 require('./graphics/viewport');
 require('./graphics/webgl');
+require('./graphics/webgl-layer');
 
 require('./isometric/diamond-iso');
 require('./isometric/isometric');
